ddl: migrate test-infra to testify for db_integration_test
