+++ ddl/serial_test.go
@@ -43,11 +43,13 @@ import (
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/store/mockstore"
+	"github.com/pingcap/tidb/table"
 	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/util/admin"
 	"github.com/pingcap/tidb/util/gcutil"
 	"github.com/pingcap/tidb/util/mock"
 	"github.com/pingcap/tidb/util/testkit"
+	"github.com/pingcap/tidb/util/testutil"
 	. "github.com/pingcap/tidb/util/testutil"
 	"github.com/tikv/client-go/v2/testutils"
 )
@@ -57,6 +59,95 @@ var _ = SerialSuites(&testSerialSuite{})
 
 // TODO(tangenta): Move all the parallel tests out of this file.
 var _ = Suite(&testIntegrationSuite7{&testIntegrationSuite{}})
+
+type testIntegrationSuite7 struct{ *testIntegrationSuite }
+
+type testIntegrationSuite struct {
+	lease   time.Duration
+	cluster testutils.Cluster
+	store   kv.Storage
+	dom     *domain.Domain
+	ctx     sessionctx.Context
+}
+
+func setupIntegrationSuite(s *testIntegrationSuite, c *C) {
+	var err error
+	s.lease = 50 * time.Millisecond
+	ddl.SetWaitTimeWhenErrorOccurred(0)
+
+	s.store, err = mockstore.NewMockStore(
+		mockstore.WithClusterInspector(func(c testutils.Cluster) {
+			mockstore.BootstrapWithSingleStore(c)
+			s.cluster = c
+		}),
+	)
+	c.Assert(err, IsNil)
+	session.SetSchemaLease(s.lease)
+	session.DisableStats4Test()
+	s.dom, err = session.BootstrapSession(s.store)
+	c.Assert(err, IsNil)
+
+	se, err := session.CreateSession4Test(s.store)
+	c.Assert(err, IsNil)
+	s.ctx = se.(sessionctx.Context)
+	_, err = se.Execute(context.Background(), "create database test_db")
+	c.Assert(err, IsNil)
+}
+
+func tearDownIntegrationSuiteTest(s *testIntegrationSuite, c *C) {
+	tk := testkit.NewTestKit(c, s.store)
+	tk.MustExec("use test")
+	r := tk.MustQuery("show tables")
+	for _, tb := range r.Rows() {
+		tableName := tb[0]
+		tk.MustExec(fmt.Sprintf("drop table %v", tableName))
+	}
+}
+
+func tearDownIntegrationSuite(s *testIntegrationSuite, c *C) {
+	s.dom.Close()
+	s.store.Close()
+}
+
+func (s *testIntegrationSuite) SetUpSuite(c *C) {
+	setupIntegrationSuite(s, c)
+}
+
+func (s *testIntegrationSuite) TearDownSuite(c *C) {
+	tearDownIntegrationSuite(s, c)
+}
+
+type testMaxTableRowIDContext struct {
+	c   *C
+	d   ddl.DDL
+	tbl table.Table
+}
+
+func newTestMaxTableRowIDContext(c *C, d ddl.DDL, tbl table.Table) *testMaxTableRowIDContext {
+	return &testMaxTableRowIDContext{
+		c:   c,
+		d:   d,
+		tbl: tbl,
+	}
+}
+
+func getMaxTableHandle(ctx *testMaxTableRowIDContext, store kv.Storage) (kv.Handle, bool) {
+	c := ctx.c
+	d := ctx.d
+	tbl := ctx.tbl
+	curVer, err := store.CurrentVersion(kv.GlobalTxnScope)
+	c.Assert(err, IsNil)
+	maxHandle, emptyTable, err := d.GetTableMaxHandle(curVer.Ver, tbl.(table.PhysicalTable))
+	c.Assert(err, IsNil)
+	return maxHandle, emptyTable
+}
+
+func checkGetMaxTableRowID(ctx *testMaxTableRowIDContext, store kv.Storage, expectEmpty bool, expectMaxHandle kv.Handle) {
+	c := ctx.c
+	maxHandle, emptyTable := getMaxTableHandle(ctx, store)
+	c.Assert(emptyTable, Equals, expectEmpty)
+	c.Assert(maxHandle, testutil.HandleEquals, expectMaxHandle)
+}
 
 type testSerialSuite struct {
 	CommonHandleSuite
